chore: rename dashbot home page file
